regroup tg handlers
