+++ src/handlers/__init__.py
@@ -1,0 +1,9 @@
+"""
+Module with all the telegram handlers.
+"""
+
+from .help_handler import help_handler
+from .error_handler import error_handler
+
+from .admin import *
+from .group import *
