+++ src/handlers/admin/__init__.py
@@ -1,0 +1,6 @@
+"""
+Module with all the telegram handlers related to admin flow.
+"""
+
+from .start_handler import start_handler
+from .menu_handler import button_handler, message_handler
