+++ src/handlers/group/__init__.py
@@ -1,0 +1,5 @@
+"""
+Module with all the telegram handlers related to group chat flow.
+"""
+
+from .group_handler import on_hashtag_message, on_new_chat_member
