+++ src/handlers/help_handler.py
@@ -1,0 +1,7 @@
+from telegram import Update
+from telegram.ext import CallbackContext
+
+from src import constants
+
+def help_handler(update: Update, _: CallbackContext):
+    update.message.reply_text(constants.help_message)
